refactor: ContentTypeConfig 클래스명 수정 및 패키지 변경
- 기존: constants.ContentTypeConfig
- 변경: http.header.ContentType
